backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
cryptography==43.0.1
email-validator==2.2.0
